refactor: rename /admin/configuration/llm to /admin/configuration/language-models
