+++ web/src/app/admin/configuration/language-models/ModelIcon.tsx
